+++ oaDX11Graphics/source/oaDX11IndexBuffer.cpp
@@ -6,7 +6,7 @@ bool oaEngineSDK::DX11IndexBuffer::init(Vector<uint32>& data)
   D3D11_BUFFER_DESC bd;
   ZeroMemory( &bd, sizeof(bd) );
   bd.Usage = D3D11_USAGE_DEFAULT;
-  bd.ByteWidth = sizeof(uint32)*data.size();
+  bd.ByteWidth = static_cast<UINT>(sizeof(uint32)*data.size());
   bd.BindFlags = D3D11_BIND_INDEX_BUFFER;
   bd.CPUAccessFlags = 0;
